# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AddIn": {
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+$",
            "title": "Email",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "admin",
              "editor",
              "operator",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ],
        "title": "AddIn",
        "type": "object"
      },
...
      "CodeIn": {
        "properties": {
          "code": {
            "maxLength": 32,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "CodeIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemberOut": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "admin",
              "editor",
              "operator",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "role"
        ],
        "title": "MemberOut",
        "type": "object"
      },
      "MemberRoleOut": {
        "additionalProperties": false,
        "properties": {
          "role": {
            "enum": [
              "owner",
              "admin",
              "editor",
              "operator",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "role"
        ],
        "title": "MemberRoleOut",
        "type": "object"
      },
...
      "RecoveryCodesOut": {
        "additionalProperties": false,
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "title": "Recovery Codes",
            "type": "array"
          },
          "state": {
            "enum": [
              "mfa_pending",
              "enroll_required",
              "active"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "recovery_codes",
          "state",
          "csrf_token"
        ],
        "title": "RecoveryCodesOut",
        "type": "object"
      },
...
      "RoleChange": {
        "properties": {
          "role": {
            "enum": [
              "owner",
              "admin",
              "editor",
              "operator",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "RoleChange",
        "type": "object"
      },
...
      "StateOut": {
        "additionalProperties": false,
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "state": {
            "enum": [
              "mfa_pending",
              "enroll_required",
              "active"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "csrf_token"
        ],
        "title": "StateOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Dewpoint API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/mfa/recovery": {
      "post": {
        "operationId": "mfa_recovery_api_v1_auth_mfa_recovery_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mfa Recovery",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/totp": {
      "post": {
        "operationId": "mfa_totp_api_v1_auth_mfa_totp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mfa Totp",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/totp/confirm": {
      "post": {
        "operationId": "confirm_api_v1_auth_mfa_totp_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/mfa/totp/reauth": {
      "post": {
        "description": "Prove the confirmed TOTP again from an active session (before changing factors).",
        "operationId": "reauth_api_v1_auth_mfa_totp_reauth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reauth",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/t/{tenant_id}/members": {
      "get": {
        "operationId": "list_members_api_v1_t__tenant_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberOut"
                  },
                  "title": "Response List Members Api V1 T  Tenant Id  Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Members",
        "tags": [
          "members"
        ]
      },
      "post": {
        "operationId": "add_api_v1_t__tenant_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add",
        "tags": [
          "members"
        ]
      }
    },
    "/api/v1/t/{tenant_id}/members/{user_id}": {
      "delete": {
        "operationId": "remove_api_v1_t__tenant_id__members__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove",
        "tags": [
          "members"
        ]
      },
      "patch": {
        "operationId": "change_api_v1_t__tenant_id__members__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change",
        "tags": [
          "members"
        ]
      }
    },
...
  }
}